requirements.txt
fastapi 
uvicorn 
google-genai 
pydantic 
python-multipart 
python-dotenv
python-docx
